social_network/pyproject.toml
[tool.poetry]
name = "social-network"
version = "0.1.0"
description = ""
authors = ["alex <me@fomin3.ru>"]
readme = "README.md"
packages = [{include = "social_network"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.89.0"
uvicorn = "^0.20.0"
tortoise-orm = {extras = ["accel"], version = "^0.19.2"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
aiohttp = {extras = ["speedups"], version = "^3.8.3"}
asyncio = "^3.4.3"
email-validator = "^1.3.0"
redis = "^4.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
